refactor: cambiar el tipo de dato Integer -> Long de la variable que almacena los boletos ocupados

+++ src/main/java/com/aldo/event_pass/repository/DetalleCompraRepository.java
@@ -21,5 +21,5 @@ public interface DetalleCompraRepository extends JpaRepository<DetalleCompra, Lo
         )
     )
     """)
-    Integer obtenerBoletosOcupados(@Param("tipoBoletoId") Long tipoBoletoId);
+    Long obtenerBoletosOcupados(@Param("tipoBoletoId") Long tipoBoletoId);
 }
